fix: rename to _vector
